refactor(health): split rollout modules
